# Replace the 32-bit scalar field with a fiat-generated implementation