@gtank 2022-03-26
Oh, noting here for later: it would be nice if `fiat_sc255_to_bytes` took a `fiat_sc255_non_montgomery_domain_field_element` as suggested by its documentation, rather than a raw `*[4]uint64`. It would save an ugly cast in the following usage

```
func (s *Scalar) bytes(out *[32]byte) []byte {
	var limbs fiat_sc255_non_montgomery_domain_field_element
	fiat_sc255_from_montgomery(&limbs, &s.s)
	fiat_sc255_to_bytes(out, (*[4]uint64)(&limbs))
	return out[:]
}
```