@FiloSottile 2022-07-31
Replaced `scReduce` with a series of operations reducing the value in limbs, following the identity `x = a + b * 2^168 + c * 2^336 mod l`. Updated the fiat-crypto code to skip some type-casting that generates worse code, probably improves performance.